% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_zcp_data.R
\name{batch_write_csv}
\alias{batch_write_csv}
\title{Batch Write Data Frames from a List to CSV Files}
\usage{
batch_write_csv(dt_list, save_dir)
}
\arguments{
\item{dt_list}{A list where each element is a \code{data.frame}. The names of the list elements
will be used as the names of the output CSV files.}

\item{save_dir}{A character string representing the directory path where the CSV files will be saved.
If the directory does not exist, it will be created.}
}
\value{
This function does not return a value. It saves CSV files to the specified directory.
}
\description{
This function takes a list of data frames and saves each data frame as a separate
CSV file in the specified directory. If the directory does not exist, it is created.
It uses the \code{purrr::walk} function to iterate over the list and save each \code{data.frame}
with its name as the CSV file name.
}
\examples{
# Example
df1 <- data.frame(a = 1:3, b = letters[1:3])
df2 <- data.frame(x = 4:6, y = letters[4:6])
batch_write_csv(dt_list=list(df1 = df1, df2 = df2), save_dir="outfile")

}
