% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_c.R
\name{dir_c}
\alias{dir_c}
\title{Create a Directory if it Doesn't Exist}
\usage{
dir_c(x = NULL)
}
\arguments{
\item{x}{A character string representing the directory path. If \code{NULL}, the
function defaults to creating a directory named "CleanedData".}
}
\description{
This function checks if the specified directory exists. If it does not exist,
it creates the directory. If no directory name is provided, it defaults to
creating a directory named "CleanedData".
}
\examples{
# Create a directory "CleanedData" (default)
dir_c()

# Create a directory "testFile"
dir_c("testFile")
}
