% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyHPPrank.R
\name{easyHPPrank}
\alias{easyHPPrank}
\title{Visualize Detected Proteins on HPP Abundance Curve}
\usage{
easyHPPrank(
  protein.ids,
  HPPms,
  pointcolor = "#CF3175",
  pointalpha = 0.5,
  pointsize = 2,
  line.cex = 0.8,
  line.color = "black",
  line.alpha = 0.2
)
}
\arguments{
\item{protein.ids}{A character vector of UniProt IDs to highlight, e.g., \code{c("P04196", "P43652", "P01871")}.}

\item{HPPms}{A \code{data.table} or \code{data.frame} containing HPP mass spectrometry data, typically from \code{getHPP(technology = "ms")$Data}.}

\item{pointcolor}{Color for highlighted points. Default is \code{"#CF3175"}.}

\item{pointalpha}{Transparency of points. Default is \code{0.5}.}

\item{pointsize}{Size of highlighted points. Default is \code{2}.}

\item{line.cex}{Line width of the loess smoothed curve. Default is \code{0.8}.}

\item{line.color}{Color of the smoothed trend line. Default is \code{"black"}.}

\item{line.alpha}{Transparency of the smoothed line. Default is \code{0.2}.}
}
\value{
A list with two elements:
\describe{
\item{\code{Figure}}{A ggplot object showing the abundance distribution and detected proteins.}
\item{\code{data}}{The input data with updated \code{detected} and \code{density} columns.}
}
}
\description{
This function visualizes user-specified proteins (by UniProt ID) against the Human Protein Atlas (HPA) blood MS abundance dataset.
It highlights detected proteins along a ranked abundance curve, with optional customization for point color, size, and loess smoothing.
}
\details{
Proteins are ranked by abundance, and user-supplied protein IDs are labeled as "Identified".
The function overlays vertical lines (by density) and points for identified proteins, along with a loess-fitted global trend.
}
\examples{

HPP_ms = ZCPstates::getHPP(dest_dir = getwd(), technology = "ms", verbose = TRUE)
plt= ZCPstates::easyHPPrank(
  protein.ids = c("P04196", "P43652", "P01871", "P10909", "P98161", 
                  "P05156", "Q06033", "P61073", "P05106", "P0DOX5", "Q8N6U8"),
  HPPms = HPP_ms$Data,
  pointcolor = "#CF3175",
  pointalpha = 0.5,
  pointsize = 2,
  line.cex = 0.8,
  line.color = "black",
  line.alpha = 0.2
)
plt$Figure


}
