% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_Extract.R
\name{Fasta_extract}
\alias{Fasta_extract}
\title{Extract Information from a FASTA File}
\usage{
Fasta_extract(
  fasta.path = "uniprotkb_proteome_UP000005640_AND_revi_2024_02_17.fasta"
)
}
\arguments{
\item{fasta.path}{A character string specifying the path to the FASTA file.}
}
\value{
A tibble data frame with three columns:
\item{Protein.Group}{UniProt ID extracted from the header.}
\item{Genes}{Gene name extracted from the header.}
\item{Description}{Gene description extracted from the header.}
}
\description{
This function extracts relevant protein information such as UniProt ID, Gene Name,
and Gene Description from a given FASTA file containing protein sequences.
}
\details{
This function reads a FASTA file, extracts the UniProt ID, Gene Name, and Gene
Description from each sequence header, and returns the extracted information in a tibble format.
The function uses regular expressions to extract these details based on the expected
format of the FASTA headers.

The FASTA header is expected to follow the format:
\code{>sp|UniProt_ID|... GN=GeneName ... Description text OS=...}

The function assumes that the header information contains at least the UniProt ID,
Gene Name, and Gene Description for proper extraction.
}
\examples{
# Example of extracting information from a FASTA file:
fasta_info <- Fasta_extract("your_fasta_file.fasta")

}
