% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumanFastaLibs}
\alias{HumanFastaLibs}
\title{Human FASTA Mapping Table}
\format{
A data.frame with columns: \code{Entry}, \code{Gene}, \code{Description}
}
\source{
Custom generated from UniProt FASTA file
}
\usage{
HumanFastaLibs
}
\description{
UniProt-Gene-Description mapping from Human UniProt FASTA.
}
\keyword{datasets}
