% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltglynetwork.R
\name{pltglynetwork}
\alias{pltglynetwork}
\title{Plot Glycan Network}
\usage{
pltglynetwork(
  dt,
  peptide.col = "Annotated Sequence",
  glycancomposition.col = "Glycan composition",
  protein.col = "Master Protein Accessions",
  outdir = "outputfile_test",
  SiteperProteinsCutoff = 4,
  edge_cex = 0.5,
  node_size = 3,
  Title = "test"
)
}
\arguments{
\item{dt}{A data.frame containing the glycoprotein data.}

\item{peptide.col}{Character. The column name in \code{dt} representing the peptide sequence. Default is "Annotated Sequence".}

\item{glycancomposition.col}{Character. The column name in \code{dt} representing the glycan composition. Default is "Glycan composition".}

\item{protein.col}{Character. The column name in \code{dt} representing the protein identifier. Default is "Master Protein Accessions".}

\item{outdir}{Character. Directory path to save output files. Default is "outputfile_test".}

\item{SiteperProteinsCutoff}{Numeric. A cutoff value for the number of glycosylation sites per protein. Default is 4.}

\item{edge_cex}{Numeric. The line width for the network edges in the plot. Default is 0.5.}

\item{node_size}{Numeric. The size of the network nodes. Default is 3.}

\item{Title}{Character. The title of the network plot. Default is "test".}
}
\value{
A ggplot object representing the glycan network plot.
}
\description{
This function processes glycoprotein data to extract glycan types and glycosylation site information, then visualizes the glycan network using ggplot2 and ggnetwork. The network comprises nodes representing glycan compositions and proteins, with edges indicating their associations.
}
\details{
The function performs the following steps:
\enumerate{
\item Selects and renames the required columns.
\item Computes the number of glycosylation sites per protein.
\item Processes the glycan composition by replacing specific substrings and extracting numeric values.
\item Classifies glycan types based on computed counts.
\item Constructs nodes and edges for the glycan network and computes a custom layout.
\item Saves intermediate data files and outputs the final network plot.
}
}
\examples{
\dontrun{
  # Example usage:
  library(ZCPstates)
  data("Glycan_network_demo", package = "ZCPstates")
  plt <- pltglynetwork(Glycan_network_demo,
                       peptide.col = "Annotated Sequence",
                       glycancomposition.col = "Glycan composition",
                       protein.col = "Master Protein Accessions",
                       Title = "Glycan Network")
  print(plt)
}

}
