% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltvenn.R
\name{pltvenn}
\alias{pltvenn}
\title{Create a Venn Diagram from Long Data Format}
\usage{
pltvenn(
  dat_long,
  group.by = "sample",
  proteincol.id = "Accession",
  fill.alpha = 0.3,
  type = c("ellipse", "circle")[1]
)
}
\arguments{
\item{dat_long}{A data frame in long format containing at least two columns:
one representing the grouping variable (\code{group.by}), and one representing the
protein identifiers (\code{proteincol.id}).}

\item{group.by}{A character string specifying the column name for the grouping variable
(default is "sample"). This column will be used to divide the data into groups.}

\item{proteincol.id}{A character string specifying the column name for protein identifiers
(default is "Accession"). This column will be used to identify unique proteins in the
Venn diagram.}

\item{fill.alpha}{A numeric value between 0 and 1 that controls the transparency of the
fill color in the Venn diagram (default is 0.3).}
}
\value{
A Venn diagram (Euler diagram) plot generated using the \code{eulerr} package. The plot
shows the intersection of proteins across the different groups.
}
\description{
This function generates a Venn diagram based on a long format data frame containing at least
two columns: one representing the grouping variable and one representing the protein identifiers.
It uses the \code{eulerr} package to create Euler diagrams, which are similar to Venn diagrams, with
customizable colors and fill alpha values.
}
\examples{
# Example
common_elements = c("a", "b", "c", "d", "e")
group1 = c(common_elements, sample(letters[6:26], 5))
group2 = c(common_elements, sample(letters[6:26], 5))
group3 = c(common_elements, sample(letters[6:26], 5))

dt = data.frame(
  id = c(group1, group2, group3),
  group = rep(c("group1", "group2", "group3"), each = length(common_elements) + 5)
)
pltvenn(dt, group.by = "group", proteincol.id = "id",type=c("ellipse","circle")[1])

}
