% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_zcp.R
\name{save_zcp}
\alias{save_zcp}
\title{Save Plots in Multiple Formats}
\usage{
save_zcp(Fig, FigName, outputfile, widths, heights, ppt = TRUE)
}
\arguments{
\item{Fig}{A \code{ggplot} object to save.}

\item{FigName}{A string representing the base name for the saved files (without extension).}

\item{outputfile}{A string representing the directory where the files will be saved. The function creates the directory if it does not exist.}

\item{widths}{A numeric value specifying the width of the saved plot.}

\item{heights}{A numeric value specifying the height of the saved plot.}

\item{ppt}{Logical. If \code{TRUE} (default), the function generates a PPT file in addition to PNG and PDF files. If \code{FALSE}, only PNG and PDF files are created.}
}
\value{
No return value. The function saves the plot files in the specified directory.
}
\description{
This function saves a given ggplot object in PDF, PNG, and optionally PPT formats.
It ensures the output directory exists and generates the files in the specified directory.
}
\examples{
\dontrun{
library(ggplot2)
#demo
example_plot <- ggplot(mtcars, aes(mpg, wt)) + geom_point()

# Save the plot
save_zcp(Fig = example_plot,
         FigName = "example_plot",
         outputfile = "outputfile",
         widths = 7,
         heights = 5,
         ppt = TRUE)
}
}
