% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniprotEnrich.demo.library.uniprot}
\alias{uniprotEnrich.demo.library.uniprot}
\title{uniprotEnrich.demo.library.uniprot: Example UniProt Database}
\format{
A data frame with N rows and M columns.
\describe{
\item{Entry}{Protein ID from UniProt.}
\item{Pathway}{Pathway annotations associated with proteins.}
\item{...}{Additional columns for GO annotations.}
}
}
\source{
\url{https://uniprot.org}
}
\usage{
uniprotEnrich.demo.library.uniprot
}
\description{
This dataset contains a sample UniProt database for proteome UP000006548.
It includes columns such as "Entry", "Pathway", and Gene Ontology annotations.
}
\keyword{datasets}
