% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniprotEnrichplot.demo.GO.CC}
\alias{uniprotEnrichplot.demo.GO.CC}
\title{uniprotEnrichplot.demo.GO.CC: Example Dataset for visualing the enrichment analysis result}
\format{
A data frame with N rows and M columns.
#' \describe{
\item{Type}{GO subterm identifiers.}
\item{counts}{Enriched pathway size.}
\item{Enrich_factor}{NES value for each pathway.}
\item{...}{Additional columns as required.}
}
}
\source{
Example data
}
\usage{
uniprotEnrichplot.demo.GO.CC
}
\description{
This dataset is derived from the analysis results of the uniprotEnrich function, specifically focusing on the CC (cellular component) subset of GO enrichment. It is intended to demonstrate the visualization capabilities of the uniprotEnrich_plot function.
}
\keyword{datasets}
